moved logout button to navbar, removed empty array from second argument in dispatch on each page
