+++ client/src/components/Nav/index.js
@@ -1,11 +1,19 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Link } from 'react-router-dom';
 import { useStoreContext } from "../../utils/GlobalState";
+import API from '../../utils/API';
 
 function Nav() {
 
   //Test nav component to read to store and return to home page
-  const [store] = useStoreContext();
+  const [store, dispatch] = useStoreContext();
+  const logOut = () =>{
+    API.logout().then(res =>{
+      console.log('RES IN LOGOUT ', res);
+      dispatch({type: 'logout'} );
+
+    }).catch(err => console.log(err));
+  }
 
   if (store.isLoggedIn) {
     return(
@@ -18,6 +26,7 @@ function Nav() {
         <Link className='nav-link' to='/create'>Create</Link>
         <Link className='nav-link' to='/recent'>Recent</Link>
         <Link className='nav-link' to='/search'>Search</Link>
+        <button className='btn btn-warning' onClick={logOut}>Logout</button>
       </nav>
     );
   } else {
