+++ client/src/pages/Calendar.js
@@ -4,14 +4,14 @@ import API from '../utils/API';
 
 function Calendar(){
 
-  const [store, dispatch] = useStoreContext();
+  const [, dispatch] = useStoreContext();
   //get user data from api and store to global context
   useEffect(()=>{
     API.getLoggedInUser().then(res =>{
       const user = res.data;
       if(res.data) { dispatch({ type: 'setuser',user: user});}
     }).catch(err=> console.log(err));
-  },[]);
+  });
 
   return(
     <h1>Calendar</h1>
