+++ client/src/pages/Search.js
@@ -3,7 +3,7 @@ import { useStoreContext } from '../utils/GlobalState'
 import API from '../utils/API';
 
 function Search(){
-  const [store, dispatch] = useStoreContext();
+  const [, dispatch] = useStoreContext();
 
   //get user data from api and store to global context
   useEffect(()=>{
@@ -11,7 +11,7 @@ function Search(){
       const user = res.data;
       if(res.data) { dispatch({ type: 'setuser',user: user});}
     }).catch(err=> console.log(err));
-  },[]);
+  });
   return(
     <h1>Search</h1>
   )
